+++ packages/api/src/__generated__/schema.ts
@@ -183,6 +183,7 @@ export type IStoreB2B = {
   firstName?: Maybe<Scalars['String']>;
   isRepresentative?: Maybe<Scalars['Boolean']>;
   lastName?: Maybe<Scalars['String']>;
+  savedPostalCode?: Maybe<Scalars['String']>;
   unitId?: Maybe<Scalars['String']>;
   unitName?: Maybe<Scalars['String']>;
   userEmail?: Maybe<Scalars['String']>;
@@ -894,6 +895,7 @@ export type StoreB2B = {
   firstName?: Maybe<Scalars['String']>;
   isRepresentative?: Maybe<Scalars['Boolean']>;
   lastName?: Maybe<Scalars['String']>;
+  savedPostalCode?: Maybe<Scalars['String']>;
   unitId?: Maybe<Scalars['String']>;
   unitName?: Maybe<Scalars['String']>;
   userEmail?: Maybe<Scalars['String']>;