+++ packages/core/src/components/cms/RenderSections.tsx
@@ -10,9 +10,9 @@ import { useUI } from '@faststore/ui'
 import type { Section } from '@vtex/client-cms'
 import dynamic from 'next/dynamic'
 import useTTI from 'src/sdk/performance/useTTI'
+import COMPONENTS from './global/Components'
 import SectionBoundary from './SectionBoundary'
 import ViewportObserver from './ViewportObserver'
-import COMPONENTS from './global/Components'
 
 interface Props {
   components?: Record<string, ComponentType<any>>