+++ packages/core/src/components/templates/ProductListingPage/ProductListingPage.tsx
@@ -26,7 +26,7 @@ export type ProductListingPageProps = {
   serverManyProductsVariables: ServerManyProductsQueryQueryVariables
   page: PLPContentType
   globalSections?: Array<{ name: string; data: any }>
-  globalSectionsSettings?: Record<string, any>
+  globalSettings?: Record<string, unknown>
 }
 
 type UseSearchParams = {
@@ -77,7 +77,7 @@ export default function ProductListingPage({
   data: server,
   serverManyProductsVariables,
   globalSections,
-  globalSectionsSettings,
+  globalSettings,
 }: ProductListingPageProps) {
   const { settings } = plpContentType
   const collection = server.collection
@@ -136,7 +136,7 @@ export default function ProductListingPage({
 
       <ProductListing
         globalSections={globalSections}
-        globalSectionsSettings={globalSectionsSettings}
+        globalSettings={globalSettings}
         page={plpContentType}
         data={server}
         serverManyProductsVariables={serverManyProductsVariables}