+++ packages/core/src/sdk/overrides/PageProvider.tsx
@@ -9,12 +9,16 @@ import type { PropsWithChildren } from 'react'
 import { createContext, useContext, useMemo } from 'react'
 import type { SearchPageContextType } from 'src/pages/s'
 
-export interface PDPContext {
+interface PageGlobalContext {
+  globalSettings?: Record<string, unknown>
+}
+
+export interface PDPContext extends PageGlobalContext {
   data?: ServerProductQueryQuery &
     ClientProductQueryQuery['product'] & { isValidating?: boolean }
 }
 
-export interface PLPContext {
+export interface PLPContext extends PageGlobalContext {
   data?: ServerCollectionPageQueryQuery &
     ClientProductGalleryQueryQuery & {
       pages: ClientManyProductsQueryQuery[]
@@ -22,7 +26,7 @@ export interface PLPContext {
   globalSectionsSettings?: Record<string, any>
 }
 
-export interface SearchPageContext {
+export interface SearchPageContext extends PageGlobalContext {
   data?: SearchPageContextType &
     ClientProductGalleryQueryQuery & {
       pages: ClientManyProductsQueryQuery[]
@@ -30,7 +34,7 @@ export interface SearchPageContext {
   globalSectionsSettings?: Record<string, any>
 }
 
-export interface DynamicContent<T> {
+export interface DynamicContent<T> extends PageGlobalContext {
   data?: T
 }
 
@@ -39,6 +43,7 @@ export interface PageProviderContextValue {
 }
 
 type PageProviderContext =
+  | PageGlobalContext
   | PDPContext
   | PLPContext
   | SearchPageContext